+++ iati/tests/resources.py
@@ -42,7 +42,9 @@ def get_test_data_path(name, version=None):
     if name[-4:] == iati.resources.FILE_DATA_EXTENSION:
         name = name[:-4]
 
-    return os.path.join(PATH_TEST_DATA, iati.resources.get_folder_name_for_version(version), '{0}'.format(name) + iati.resources.FILE_DATA_EXTENSION)
+    relative_path = os.path.join(PATH_TEST_DATA, iati.resources.get_folder_name_for_version(version), '{0}'.format(name) + iati.resources.FILE_DATA_EXTENSION)
+
+    return iati.resources.resource_filename(relative_path)
 
 
 def get_test_data_paths_in_folder(folder_name, version=None):
@@ -71,7 +73,7 @@ def get_test_data_paths_in_folder(folder_name, version=None):
             paths.append(os.path.join(base_folder, file_name))
 
     # de-resource the file-names so that they're not duplicated
-    deresourced_paths = [path[path.find(root_folder):] for path in paths]
+    deresourced_paths = [iati.resources.resource_filename(path[path.find(root_folder):]) for path in paths]
 
     return deresourced_paths
 
@@ -101,11 +103,11 @@ def get_test_ruleset_path(name, version=None):
     return os.path.join(PATH_TEST_DATA, iati.resources.get_folder_name_for_version(version), 'rulesets/{0}'.format(name) + iati.resources.FILE_RULESET_EXTENSION)
 
 
-def load_as_dataset(file_path):
+def load_as_dataset(relative_path):
     """Load a specified test data file as a Dataset.
 
     Args:
-        file_path (str): The path of the file, relative to the root test data folder. Folders should be separated by a forward-slash (`/`).
+        relative_path (str): The path of the file, relative to the root test data folder. Folders should be separated by a forward-slash (`/`).
 
     Returns:
         dataset: A Dataset containing the contents of the file at the specified location.
@@ -114,17 +116,21 @@ def load_as_dataset(file_path):
         iati.exceptions.ValidationError: If the provided XML does not conform to the IATI standard.
 
     """
-    return iati.resources.load_as_dataset(iati.tests.resources.get_test_data_path(file_path))
+    path = iati.tests.resources.get_test_data_path(relative_path)
+
+    return iati.utilities.load_as_dataset(path)
 
 
-def load_as_string(file_path):
+def load_as_string(relative_path):
     """Load a specified test data file as a string.
 
     Args:
-        file_path (str): The path of the file, relative to the root test data folder. Folders should be separated by a forward-slash (`/`).
+        relative_path (str): The path of the file, relative to the root test data folder. Folders should be separated by a forward-slash (`/`).
 
     Returns:
         str (python3) / unicode (python2): The contents of the file at the specified location.
 
     """
-    return iati.resources.load_as_string(iati.tests.resources.get_test_data_path(file_path))
+    path = iati.tests.resources.get_test_data_path(relative_path)
+
+    return iati.utilities.load_as_string(path)