+++ iati/tests/test_codelists.py
@@ -65,7 +65,7 @@ class TestCodelists(object):
     def test_codelist_define_from_xml(self, name_to_set):
         """Check that a Codelist can be generated from an XML codelist definition."""
         path = iati.resources.get_codelist_path('FlowType')
-        xml_str = iati.resources.load_as_string(path)
+        xml_str = iati.utilities.load_as_string(path)
         codelist = iati.Codelist(name_to_set, xml=xml_str)
 
         code_names = ['ODA', 'OOF', 'Private grants', 'Private Market', 'Non flow', 'Other flows']
@@ -81,7 +81,7 @@ class TestCodelists(object):
         """Check that a Codelist can be generated from an XML codelist definition."""
         codelist_name = 'BudgetType'
         path = iati.resources.get_codelist_path(codelist_name)
-        xml_str = iati.resources.load_as_string(path)
+        xml_str = iati.utilities.load_as_string(path)
         codelist = iati.Codelist(codelist_name, xml=xml_str)
 
         assert codelist.complete is True
@@ -90,7 +90,7 @@ class TestCodelists(object):
         """Check that a Codelist can be generated from an XML codelist definition."""
         codelist_name = 'Country'
         path = iati.resources.get_codelist_path(codelist_name)
-        xml_str = iati.resources.load_as_string(path)
+        xml_str = iati.utilities.load_as_string(path)
         codelist = iati.Codelist(codelist_name, xml=xml_str)
 
         assert codelist.complete is False