+++ iati/tests/utilities.py
@@ -1,29 +1,12 @@
-"""A module containing utility constants and functions for tests.
-
-A large number of constants containing example file content are contained. These constants are named from left to right, with general properties first, then leading into more specific information. These names indicate what they are used for.
-
-Example:
-    To load a file into a string::
-
-        name_of_file = 'a-file-name-without-the-extension'
-        CONSTANT_NAME = load_as_string(name_of_file)
-
-Note:
-    The current method of managing test data is known to be sub-optimal. Suggestions for better methods that satisfy requirements are appreciated!
-
-Todo:
-    Add versions of constants that are valid for differing schema versions.
-
-"""
+"""A module containing utility constants and functions for tests."""
 import decimal
-from lxml import etree
 import iati.resources
 import iati.constants
 import iati.tests.resources
 
 
 # This will need updating once test data forcing XML format is fixed
-RULESET_FOR_TESTING = iati.Ruleset(iati.resources.load_as_string(iati.resources.get_ruleset_path('ruleset_for_tests')))
+RULESET_FOR_TESTING = iati.Ruleset(iati.utilities.load_as_string(iati.resources.get_ruleset_path('ruleset_for_tests')))
 """A working Ruleset based on the Standard Ruleset."""
 
 
@@ -34,11 +17,11 @@ SCHEMA_ORGANISATION_NAME_VALID = 'iati-organisations-schema'
 SCHEMA_NAME_VALID = 'iati-activities-schema'
 """A string containing a valid Schema name."""
 
-XML_TREE_VALID = etree.fromstring(iati.tests.resources.load_as_string('valid_not_iati'))
+XML_TREE_VALID = iati.utilities.load_as_tree(iati.tests.resources.get_test_data_path('valid_not_iati'))
 """An etree that is valid XML but not IATI XML."""
-XML_TREE_VALID_IATI = etree.fromstring(iati.tests.resources.load_as_string('valid_iati'))
+XML_TREE_VALID_IATI = iati.utilities.load_as_tree(iati.tests.resources.get_test_data_path('valid_iati'))
 """A valid IATI etree."""
-XML_TREE_VALID_IATI_INVALID_CODE = etree.fromstring(iati.tests.resources.load_as_string('valid_iati_invalid_code'))
+XML_TREE_VALID_IATI_INVALID_CODE = iati.utilities.load_as_tree(iati.tests.resources.get_test_data_path('valid_iati_invalid_code'))
 """A valid IATI etree that has an invalid Code value."""